Move tests/python_direct to tests/python/direct
