+++ tests/python/direct/test_import.py
